requirements.txt
altair==4.2.0
docx2txt==0.8
matplotlib==3.5.1
numpy==1.22.2
pandas==1.4.1
pdfplumber==0.6.0
Pillow==9.0.1
plotly==5.6.0
PyPDF2==1.26.0
seaborn==0.11.2
streamlit==1.6.0
wordcloud==1.8.1
